pyproject.toml
[project]
name = "tips-portal"
version = "0.1.0"
authors = [
    {name = "Carlos Moreno", email = "omorenodovale@gmail.com"},
]
description = "Portal with quick command tips."
dynamic = ["readme"]
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.11",
    "flask>=3.1.1",
    "flask-pymongo>=3.0.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "taskipy",
    "bandit",
]

[tool.setuptools.dynamic]
readme = {file = ["README.md"]}

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['C90', 'I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.bandit]
exclude_dirs = [".venv"]
skips = ["B101"]

[tool.taskipy.tasks]
lint = 'uv run ruff check --fix portal tests'
format = 'uv run ruff format portal tests'
security = 'uv run bandit -c pyproject.toml -r .'
pre_test = 'task lint'
test = """
    uv run pytest --cov=portal -vv
    uv run coverage xml
    uv run coverage html
    """
clean = """
    find . -depth -maxdepth 5 -name '*.pyc' -exec rm -rf {} \\;
	find . -depth -maxdepth 5 -name '__pycache__' -exec rm -rf {} \\;
	find . -depth -maxdepth 5 -name 'Thumbs.db' -exec rm -rf {} \\;
	find . -depth -maxdepth 5 -name '*~' -exec rm -rf {} \\;
    rm -rf .pytest_cache
    rm -rf .ruff_cache
    rm -rf build
    rm -rf dist
    rm -rf portal.egg-info
    rm -rf htmlcov
    rm -rf coverage.xml
    rm -rf .coverage
    """
